:recycle: Rename folder core to init
