+++ internal/actor/system/init/actor.go
@@ -1,10 +1,10 @@
-package core
+package init
 
 import (
 	"github.com/asynkron/protoactor-go/actor"
 
-	"github.com/ccamel/playground-protoactor.go/internal/actor/system/core/sys"
-	"github.com/ccamel/playground-protoactor.go/internal/actor/system/core/usr"
+	"github.com/ccamel/playground-protoactor.go/internal/actor/system/init/sys"
+	"github.com/ccamel/playground-protoactor.go/internal/actor/system/init/usr"
 )
 
 type Actor struct{}
